- refactor tests - remove spotless - reformat files
