+++ library/src/main/java/com/patryklikus/IFJV/library/util/GsonUtils.java
@@ -5,6 +5,7 @@ import com.google.gson.JsonArray;
 import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import com.google.gson.JsonPrimitive;
+
 import javax.annotation.Nullable;
 
 public class GsonUtils {
