+++ library/src/test/java/com/patryklikus/IFJV/library/validator/JsonBooleanValidatorTest.java
@@ -1,16 +1,15 @@
 /* Copyright Patryk Likus All Rights Reserved. */
 package com.patryklikus.IFJV.library.validator;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 
-@DisplayName("JsonBooleanValidatorImpl")
-public class JsonBooleanValidatorTest {
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+class JsonBooleanValidatorTest {
     private JsonBooleanValidator jsonBooleanValidator;
 
     @BeforeEach
