+++ library/src/test/java/com/patryklikus/IFJV/library/validator/JsonObjectValidatorTest.java
@@ -1,16 +1,10 @@
 /* Copyright Patryk Likus All Rights Reserved. */
 package com.patryklikus.IFJV.library.validator;
 
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.mockito.Mockito.when;
-
 import com.patryklikus.IFJV.library.schema.model.BooleanSchema;
 import com.patryklikus.IFJV.library.schema.model.IntegerSchema;
 import com.patryklikus.IFJV.library.schema.model.JsonSchema;
 import com.patryklikus.IFJV.library.schema.model.ObjectSchema;
-import java.util.Map;
-import java.util.Set;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
@@ -18,7 +12,13 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
-@DisplayName("JsonArrayValidatorImpl")
+import java.util.Map;
+import java.util.Set;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+import static org.mockito.Mockito.when;
+
 @ExtendWith(MockitoExtension.class)
 class JsonObjectValidatorTest {
     @Mock
